Refactor CoonectionGrid
