+++ web/src/components/ConnectionGrid/type.ts
@@ -1,0 +1,3 @@
+import { ConnectionArgs } from "../../graphql/type/type";
+
+export type RefetchFunction<TOrderField extends string> = (args?: ConnectionArgs<TOrderField>) => void;
